# speeds up tablet mgmt iterator