--- a/core/src/main/java/org/apache/accumulo/core/metadata/schema/TabletMetadata.java
+++ b/core/src/main/java/org/apache/accumulo/core/metadata/schema/TabletMetadata.java
@@ -157,6 +157,8 @@ public class TabletMetadata {
     this.userCompactionsRequested = tmBuilder.userCompactionsRequested.build();
     this.unSplittableMetadata = tmBuilder.unSplittableMetadata;
     this.fileSize = Suppliers.memoize(() -> {
+      // This code was using a java stream. While profiling SplitMillionIT, the stream was showing
+      // up as hot when scanning 1 million tablets. Converted to a for loop to improve performance.
       long sum = 0;
       for (var dfv : files.values()) {
         sum += dfv.getSize();