Based on profiling tablet servers while running SplitMillionIT, made the following changes related to TabletManagementIterator

 * Made TabletMgmtIterator extend WholeRowIterator instead of stack on top of it.  This avoids encoding->decoding->encoding rows on the iterator stack.
 * Avoided lookuping and parsing multiple table props for each tablet inorder to determine if it needs split. Moved this be done once per table.
 * Changed a stream to compute file size sum into a loop.