@cshannon 2024-05-17
Did you compare the test time before and after the changes? Just curious on the performance difference.