@dlmarion 2024-05-17
> Changed a stream to compute file size sum into a loop.

This is the Nth time where we are undo-ing stream operations for performance reasons...