@keith-turner 2024-05-17
> Did you compare the test time before and after the changes? Just curious on the performance difference.

Yeah I did see a significant performance increase for scanning 1 million tablets.  Sampling the logs where the manager logs how long a full scan of the metadata table took saw it was about 1.3 to 1.4 times faster.  Suspect the change to wholerow iterator contributes most of that, but not sure.   For each individual change the way I know it has an impact is that the code changed will be much less prominent in the profiling data.