@keith-turner 2024-05-17
> I wonder if there is a way to make Tablet hosting faster

Have a possible way, not complete sure about it.  Going to open an issue