@dlmarion 2024-05-17
MANAGER_TABLET_GROUP_WATCHER_INTERVAL defaults to 5s in MiniAccumuloClusterImpl. Can override in the test.